Changed Root Path to product#index from home#index and removed home controller
