+++ app/views/admin/categories/edit.html.erb
@@ -1,7 +1,7 @@
 <%= stylesheet_link_tag "new_user.css", "data-turbo-track": "reload" %>
 
 <div class="back-btn">
-    <%= link_to '<-Back', categories_path %>
+    <%= link_to '<-Back', admin_categories_path %>
 </div>
 
 <div class="h-100">
