+++ app/views/admin/categories/index.html.erb
@@ -57,13 +57,13 @@
                     <%= category.description %>
                 </td>                                                          
                 <td>
-                    <%= link_to category_path(category),class:"btn btn-outline-primary" do %>
+                    <%= link_to admin_category_path(category),class:"btn btn-outline-primary" do %>
                         <i class="bi bi-eye-fill" data-bs-toggle="tooltip" data-bs-placement="top" data-bs-title="Show Category"></i>
                     <% end %>
-                    <%= link_to edit_category_path(category),class:"btn btn-outline-dark"  do %>
+                    <%= link_to edit_admin_category_path(category),class:"btn btn-outline-dark"  do %>
                         <i class="bi bi-pen-fill" data-bs-toggle="tooltip" data-bs-placement="top" data-bs-title="Edit Category"></i>
                     <% end %>
-                    <%= link_to category_path(category), data: { turbo_method: :delete, turbo_confirm: 'Are you sure to delete the Category?' },class:"btn btn-outline-danger" do %>
+                    <%= link_to admin_category_path(category), data: { turbo_method: :delete, turbo_confirm: 'Are you sure to delete the Category?' },class:"btn btn-outline-danger" do %>
                         <i class="bi bi-trash-fill" data-bs-toggle="tooltip" data-bs-placement="top" data-bs-title="Delete Category"></i>
                     <% end %>
                    
@@ -76,5 +76,5 @@
 </table>
 
 <div align="center" class="mt-4">
-    <%= link_to 'Add Another', new_category_path,class:"btn btn-outline-success" %>
+    <%= link_to 'Add Another', new_admin_category_path,class:"btn btn-outline-success" %>
 </div>
