+++ app/views/admin/categories/show.html.erb
@@ -13,7 +13,7 @@
         <span class = "fw-light fs-5"><%= @category.description %></span>
     </div>
     <div class = "mt-5">
-        <%= link_to "Edit", edit_category_path(@category), class: "btn btn-primary btn-lg"%>
+        <%= link_to "Edit", edit_admin_category_path(@category), class: "btn btn-primary btn-lg"%>
     </div>
     </div>
