--- app/views/home/categories.html.erb
@@ -1,85 +1,0 @@
-<%= stylesheet_link_tag 'home.css' %>
-<div class="main row px-3 bg-light">
-  <div class="left col-md-2">
-    <!-- <div id="html1"> -->
-        <!-- <ul> -->
-          <!-- <li>Root node 1</li> -->
-          <!-- <li>Root node 2</li> -->
-          <!-- <li class="jstree-open" id="node_1">Root -->
-            <!-- <ul> -->
-              <!-- <li> -->
-                <!-- <a href="#" class="jstree-clicked">Child</a> -->
-              <!-- </li> -->
-            <!-- </ul> -->
-          <!-- </li> -->
-        <!-- </ul> -->
-      <!--  -->
-    <!-- </div> -->
-    <div class="">
-      <div class="my-4 fs-4 fw-medium">
-        Categories
-      </div>
-      <% @categories.each_with_index do |category,i| %>
-        <% if !category.subcategories.blank? %>
-            <div>
-              <div class="accordion" id="accordionExample">
-                  <div class="accordion-item">
-                      <div class="accordion-header d-flex rounded-1">
-                        <div class="my-auto ms-3 w-75">
-                          <%= link_to home_category_path(category.id) do %>
-                            <%= category.name %>
-                          <% end %>
-                        </div>
-                        <div>
-                          <button class="accordion-button collapsed" type="button" data-bs-toggle="collapse" data-bs-target="#collapse<%= i %>" aria-expanded="true" aria-controls="collapseOne">
-                            
-                        </button>
-                        </div>
-                      </div>
-                      <div id="collapse<%= i %>" class="accordion-collapse collapse" data-bs-parent="#accordionExample">
-                          <% category.subcategories.pluck(:name,:id).each do |item| %>
-                              <div class="accordion-body">
-                              <%= link_to home_category_path(item[1]) do %>
-                                  <%= item[0] %>
-                              <% end %>
-                          </div>
-                          <% end %> 
-                          
-                      </div>
-                  </div>
-              </div>
-          </div>
-          <% elsif !category.parent %>
-          <div class="w-100 bg-white p-3 fs-6 rounded-1 border">
-            <%= link_to home_category_path(category.id) do %>
-                <%= category.name %>
-            <% end %>
-          </div>
-          <% end %>
-      <% end %>
-    </div>
-  </div>
-<div class="right col-md-10">
-  <div class="electronics bg-white py-4 my-3 d-flex flex-column">
-    <div class="heading fs-3 ps-5 pb-2">
-      Products of <%= @name %> Categories
-    </div>
-    <div class="main row mx-5">
-      <% @products.each do |item| %>
-        <%= link_to product_path(item.id), class:"col-md-3" do %>
-          <div class="home-card  gap d-flex  align-items-center justify-content-center">
-            <div class="d-flex flex-column justify-content-center">
-              <%= image_tag item.images.first, class:"img-fluid h-300"%>
-              <div class="text-secondary text-center mt-2"><%= item.name %></div>
-              <p class="fs-5 text-semibold text-center">
-                &#8377;<%= item.price %>
-              </p>
-            </div>
-          </div>
-        <% end %>
-      <% end %>
-    </div>
-  </div>
-</div>
-</div>
-
