+++ app/views/shared/_nav.html.erb
@@ -51,8 +51,8 @@
             <%= link_to 'Add User +', new_user_registration_path, class: "dropdown-item" %>
             <%= link_to 'Roles', roles_path, class: "dropdown-item" %>
             <%= link_to 'Users', users_path, class: "dropdown-item" %>
-            <%= link_to 'Categories', categories_path, class: "dropdown-item" %>
-           <%= link_to 'Products', products_path, class: "dropdown-item" %>
+            <%= link_to 'Categories', admin_categories_path, class: "dropdown-item" %>
+           <%= link_to 'Products', admin_products_path, class: "dropdown-item" %>
            <%= link_to 'Stocks', product_stocks_path, class: "dropdown-item" %>
           </ul>
         </div>
